Initial refactors and code clean up
