+++ router/userRoutes.js
@@ -7,24 +7,7 @@ const generateToken = require('../helpers/generate_token')
 const isAuthenticated = require('../Middlewares/auth')
 const hideDetails = require('../helpers/hideDetails')
 const removeTasks = require('../helpers/removeTask')
-const multer = require('multer')
-
-
-let validType = ['JPG', 'JPEG', 'PNG']
-
-const upload = multer({
-    limits: {
-        fileSize: 1000000
-    },
-    fileFilter(req, file, cb) {
-      
-        let isValid = validType.filter((type) => {
-            return file.originalname.toUpperCase().endsWith(type); 
-        })
-        isValid[0] !== undefined ? cb(null, true) : cb(new Error('upload a valid image file'))
-        
-    }
-})
+const upload = require('../config/multer')
 
 
 router.post('/user', async (req, res) => {
@@ -41,9 +24,7 @@ router.post('/user', async (req, res) => {
     try {
 
         const newUser = await user.save();
-
         await generateToken(newUser)
-
         return res.status(201).send({newUser})
         
     }
@@ -99,7 +80,6 @@ router.patch('/user/me', isAuthenticated, async (req, res) => {
         return res.status(400).send('Bad update field')
     }
 
-
     try {
         let user = await User.findByIdAndUpdate({_id: req.user._id}, req.body, {new: true, runValidators: true})
         hideDetails(user)
@@ -139,33 +119,46 @@ router.post('/user/logout', isAuthenticated, async (req, res) => {
 
 
  router.post('/user/dp/me', isAuthenticated, upload.single('avatar'), async (req, res) => {
-    const buffer = await sharp(req.file.buffer).png().resize({width: 200, height: 200}).toBuffer();
-    req.user.displayPicture = buffer;
-    await req.user.save();
-    res.send('done')
+
+    try {
+        const buffer = await sharp(req.file.buffer).png().resize({width: 200, height: 200}).toBuffer();
+        req.user.displayPicture = buffer;
+        await req.user.save();
+        return res.send('Picture uploaded successfully')
+    } catch (error) {
+
+        return res.status(500).send('Soemthing went wrong')
+    }
+
  }, (error, req, res, next) => {
      return res.status(400).send({error: error.message})
  })
 
  router.delete('/user/avatar/me', isAuthenticated, async (req, res) => {
-    req.user.displayPicture = undefined;
-    await req.user.save()
-    res.send('done');
+
+     try {
+        req.user.displayPicture = undefined;
+        await req.user.save()
+        return res.send('Successfully deleted')
+     } catch (error) {
+         return res.status(500).send('Sorry, unable to delete picture. Something went wrong!')
+     }
+
  })
 
  router.get('/user/dp/me', isAuthenticated, async (req, res) => {
+
      try {
          let user = await User.findById(req.user._id);
-        
          if(user && user.displayPicture) {
              res.set('Content-Type', 'image/png')
              return res.send(user.displayPicture)
          }
-         
          throw new Error('Unable to find user display picture')
      } catch (e) {
         res.status(400).send({error: e})
      }
+
  })
